Move fixtures def to conftest.py
