+++ tests/conftest.py
@@ -2,6 +2,10 @@ import aiohttp
 import pytest
 
 from fpl import FPL
+from fpl.models import Fixture, H2HLeague, User
+from tests.test_fixture import fixture_data
+from tests.test_h2h_league import h2h_league_data
+from tests.test_user import user_data
 
 
 @pytest.fixture()
@@ -57,3 +61,20 @@ async def team():
     await session.close()
 
 
+@pytest.fixture()
+def fixture():
+    return Fixture(fixture_data)
+
+
+@pytest.fixture()
+async def h2h_league():
+    session = aiohttp.ClientSession()
+    yield H2HLeague(h2h_league_data, session)
+    await session.close()
+
+
+@pytest.fixture()
+async def user():
+    session = aiohttp.ClientSession()
+    yield User(user_data, session)
+    await session.close()
